data_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voyage-a-rome"
version = "0.1.0"
description = "Data pipeline for ROME métier data"
requires-python = ">=3.8"
dependencies = [
  "requests>=2.28",
  "PyYAML>=6.0",
  "rich>=13.0",
  "click>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["explo_rome*"]
